+++ lib/src/async_database_exception.dart
@@ -1,0 +1,8 @@
+class AsyncDatabaseException implements Exception {
+  final String message;
+
+  const AsyncDatabaseException(this.message);
+
+  @override
+  String toString() => message;
+}